feat: Phase 5 Task 6.1 - CLI分割によるPresentation層リファクタリング
## 変更内容
- src/presentation/cli.ts を903行から96行に削減（89%削減）
- コマンド登録ロジックを8つのファイルに分離:
  - jira/register.ts
  - confluence/register.ts
  - phase/register.ts
  - spec/register.ts
  - workflow/register.ts
  - config/register.ts
  - init/register.ts
  - multi-repo/register.ts

## 品質保証
- Type Check: PASS
- Lint: PASS
- Test: 1017 tests passed, 5 skipped

## アーキテクチャ改善
cli.tsはコマンド登録のみを保持し、各コマンドグループのロジックは
./commands/{group}/register.tsに委譲することで、単一責任の原則を適用。
